refactor(clients): asset client merged into query client
