+++ x-pack/platform/plugins/shared/streams/server/lib/streams/assets/query/query_client.ts
@@ -5,19 +5,170 @@
  * 2.0.
  */
 
+import type { QueryDslQueryContainer } from '@elastic/elasticsearch/lib/api/types';
 import { isBoom } from '@hapi/boom';
 import type { RulesClient } from '@kbn/alerting-plugin/server';
-import type { Logger } from '@kbn/core/server';
-import type { StreamQuery } from '@kbn/streams-schema';
+import type { Logger, SavedObjectsClientContract } from '@kbn/core/server';
+import type { IStorageClient } from '@kbn/storage-adapter';
+import type { FeatureType, StreamQuery } from '@kbn/streams-schema';
 import { buildEsqlQuery } from '@kbn/streams-schema';
 import { isEqual, map, partition } from 'lodash';
+import objectHash from 'object-hash';
 import pLimit from 'p-limit';
-import type { QueryLink } from '../../../../../common/assets';
+import type {
+  Asset,
+  AssetLink,
+  AssetLinkRequest,
+  AssetUnlinkRequest,
+  QueryLink,
+} from '../../../../../common/assets';
 import type { EsqlRuleParams } from '../../../rules/esql/types';
-import type { AssetClient } from '../asset_client';
-import { getAssetLinkUuid } from '../asset_client';
-import { ASSET_ID, ASSET_TYPE } from '../fields';
+import { AssetNotFoundError } from '../../errors/asset_not_found_error';
+import {
+  ASSET_ID,
+  ASSET_TYPE,
+  ASSET_UUID,
+  QUERY_EVIDENCE,
+  QUERY_FEATURE_FILTER,
+  QUERY_FEATURE_NAME,
+  QUERY_FEATURE_TYPE,
+  QUERY_KQL_BODY,
+  QUERY_SEVERITY_SCORE,
+  QUERY_TITLE,
+  STREAM_NAME,
+} from '../fields';
+import type { AssetStorageSettings } from '../storage_settings';
 import { getRuleIdFromQueryLink } from './helpers/query';
+
+type TermQueryFieldValue = string | boolean | number | null;
+
+interface TermQueryOpts {
+  queryEmptyString: boolean;
+}
+
+function termQuery<T extends string>(
+  field: T,
+  value: TermQueryFieldValue | undefined,
+  opts: TermQueryOpts = { queryEmptyString: true }
+): QueryDslQueryContainer[] {
+  if (value === null || value === undefined || (!opts.queryEmptyString && value === '')) {
+    return [];
+  }
+
+  return [{ term: { [field]: value } }];
+}
+
+function termsQuery<T extends string>(
+  field: T,
+  values: Array<TermQueryFieldValue | undefined> | null | undefined
+): QueryDslQueryContainer[] {
+  if (values === null || values === undefined || values.length === 0) {
+    return [];
+  }
+
+  const filteredValues = values.filter(
+    (value) => value !== undefined
+  ) as unknown as TermQueryFieldValue[];
+
+  return [{ terms: { [field]: filteredValues } }];
+}
+
+function wildcardQuery<T extends string>(
+  field: T,
+  value: TermQueryFieldValue | undefined
+): QueryDslQueryContainer[] {
+  if (value === null || value === undefined || value === '') {
+    return [];
+  }
+
+  return [{ wildcard: { [field]: { value: `*${value}*`, case_insensitive: true } } }];
+}
+
+export function getQueryLinkUuid(name: string, asset: Pick<AssetLink, 'asset.id' | 'asset.type'>) {
+  return objectHash({
+    [STREAM_NAME]: name,
+    [ASSET_ID]: asset[ASSET_ID],
+    [ASSET_TYPE]: asset[ASSET_TYPE],
+  });
+}
+
+function toQueryLink<TQueryLink extends AssetLinkRequest>(
+  name: string,
+  asset: TQueryLink
+): TQueryLink & { [ASSET_UUID]: string } {
+  return {
+    ...asset,
+    [ASSET_UUID]: getQueryLinkUuid(name, asset),
+  };
+}
+
+type StoredQueryLink = Omit<QueryLink, 'query'> & {
+  [QUERY_TITLE]: string;
+  [QUERY_KQL_BODY]: string;
+  [QUERY_SEVERITY_SCORE]?: number;
+};
+
+export type StoredAssetLink = StoredQueryLink & {
+  [STREAM_NAME]: string;
+};
+
+interface QueryBulkIndexOperation {
+  index: { asset: AssetLinkRequest };
+}
+interface QueryBulkDeleteOperation {
+  delete: { asset: AssetUnlinkRequest };
+}
+
+export type QueryBulkOperation = QueryBulkIndexOperation | QueryBulkDeleteOperation;
+
+function fromStorage(link: StoredAssetLink): QueryLink {
+  const storedQueryLink: StoredQueryLink & {
+    [QUERY_FEATURE_NAME]: string;
+    [QUERY_FEATURE_FILTER]: string;
+    [QUERY_FEATURE_TYPE]: FeatureType;
+    [QUERY_EVIDENCE]?: string[];
+  } = link as StoredAssetLink & {
+    [QUERY_FEATURE_NAME]: string;
+    [QUERY_FEATURE_FILTER]: string;
+    [QUERY_FEATURE_TYPE]: FeatureType;
+    [QUERY_EVIDENCE]?: string[];
+  };
+  return {
+    ...storedQueryLink,
+    query: {
+      id: storedQueryLink[ASSET_ID],
+      title: storedQueryLink[QUERY_TITLE],
+      kql: {
+        query: storedQueryLink[QUERY_KQL_BODY],
+      },
+      feature: storedQueryLink[QUERY_FEATURE_NAME]
+        ? {
+            name: storedQueryLink[QUERY_FEATURE_NAME],
+            filter: JSON.parse(storedQueryLink[QUERY_FEATURE_FILTER]),
+            type: storedQueryLink[QUERY_FEATURE_TYPE] ?? 'system',
+          }
+        : undefined,
+      severity_score: storedQueryLink[QUERY_SEVERITY_SCORE],
+      evidence: storedQueryLink[QUERY_EVIDENCE],
+    },
+  } satisfies QueryLink;
+}
+
+function toStorage(name: string, request: AssetLinkRequest): StoredAssetLink {
+  const link = toQueryLink(name, request);
+  const { query, ...rest } = link;
+  return {
+    ...rest,
+    [STREAM_NAME]: name,
+    [QUERY_TITLE]: query.title,
+    [QUERY_KQL_BODY]: query.kql.query,
+    [QUERY_FEATURE_NAME]: query.feature ? query.feature.name : '',
+    [QUERY_FEATURE_FILTER]: query.feature ? JSON.stringify(query.feature.filter) : '',
+    [QUERY_FEATURE_TYPE]: query.feature ? query.feature.type : '',
+    [QUERY_SEVERITY_SCORE]: query.severity_score,
+    [QUERY_EVIDENCE]: query.evidence,
+  } as unknown as StoredAssetLink;
+}
 
 function hasBreakingChange(currentQuery: StreamQuery, nextQuery: StreamQuery): boolean {
   return (
@@ -26,9 +177,9 @@ function hasBreakingChange(currentQuery: StreamQuery, nextQuery: StreamQuery): b
   );
 }
 
-function toQueryLink(query: StreamQuery, stream: string): QueryLink {
+function toQueryLinkFromQuery(query: StreamQuery, stream: string): QueryLink {
   return {
-    'asset.uuid': getAssetLinkUuid(stream, { 'asset.type': 'query', 'asset.id': query.id }),
+    'asset.uuid': getQueryLinkUuid(stream, { 'asset.type': 'query', 'asset.id': query.id }),
     'asset.type': 'query',
     'asset.id': query.id,
     query,
@@ -38,12 +189,196 @@ function toQueryLink(query: StreamQuery, stream: string): QueryLink {
 export class QueryClient {
   constructor(
     private readonly dependencies: {
-      assetClient: AssetClient;
+      storageClient: IStorageClient<AssetStorageSettings, StoredAssetLink>;
+      soClient: SavedObjectsClientContract;
       rulesClient: RulesClient;
       logger: Logger;
     },
     private readonly isSignificantEventsEnabled: boolean = false
   ) {}
+
+  // ==================== Storage Operations ====================
+
+  async linkQuery(name: string, link: AssetLinkRequest): Promise<QueryLink> {
+    const document = toStorage(name, link);
+
+    await this.dependencies.storageClient.index({
+      id: document[ASSET_UUID],
+      document,
+    });
+
+    return toQueryLink(name, link);
+  }
+
+  async syncQueryList(
+    name: string,
+    links: AssetLinkRequest[]
+  ): Promise<{ deleted: QueryLink[]; indexed: QueryLink[] }> {
+    const assetsResponse = await this.dependencies.storageClient.search({
+      size: 10_000,
+      track_total_hits: false,
+      query: {
+        bool: {
+          filter: [...termQuery(STREAM_NAME, name), ...termQuery(ASSET_TYPE, 'query')],
+        },
+      },
+    });
+
+    const existingQueryLinks = assetsResponse.hits.hits.map((hit) => {
+      return fromStorage(hit._source);
+    });
+
+    const nextQueryLinks = links.map((link) => {
+      return toQueryLink(name, link);
+    });
+
+    const nextIds = new Set(nextQueryLinks.map((link) => link[ASSET_UUID]));
+    const queryLinksDeleted = existingQueryLinks.filter((link) => !nextIds.has(link[ASSET_UUID]));
+
+    const operations: QueryBulkOperation[] = [
+      ...queryLinksDeleted.map((asset) => ({ delete: { asset } })),
+      ...nextQueryLinks.map((asset) => ({ index: { asset } })),
+    ];
+
+    if (operations.length) {
+      await this.bulkStorage(name, operations);
+    }
+
+    return {
+      deleted: queryLinksDeleted,
+      indexed: nextQueryLinks,
+    };
+  }
+
+  async unlinkQuery(name: string, asset: AssetUnlinkRequest): Promise<void> {
+    const id = getQueryLinkUuid(name, asset);
+
+    const { result } = await this.dependencies.storageClient.delete({ id });
+    if (result === 'not_found') {
+      throw new AssetNotFoundError(`${asset[ASSET_TYPE]} not found`);
+    }
+  }
+
+  async clean() {
+    await this.dependencies.storageClient.clean();
+  }
+
+  async getQueryLinks(names: string[]): Promise<Record<string, QueryLink[]>> {
+    const filters = [...termsQuery(STREAM_NAME, names), ...termQuery(ASSET_TYPE, 'query')];
+
+    const assetsResponse = await this.dependencies.storageClient.search({
+      size: 10_000,
+      track_total_hits: false,
+      query: {
+        bool: {
+          filter: filters,
+        },
+      },
+    });
+
+    const queriesPerName = names.reduce((acc, name) => {
+      acc[name] = [];
+      return acc;
+    }, {} as Record<string, QueryLink[]>);
+
+    assetsResponse.hits.hits.forEach((hit) => {
+      const name = hit._source[STREAM_NAME];
+      const asset = fromStorage(hit._source);
+      queriesPerName[name].push(asset);
+    });
+
+    return queriesPerName;
+  }
+
+  async bulkGetByIds(name: string, ids: string[]) {
+    const assetsResponse = await this.dependencies.storageClient.search({
+      size: 10_000,
+      track_total_hits: false,
+      query: {
+        bool: {
+          filter: [
+            ...termQuery(STREAM_NAME, name),
+            ...termQuery(ASSET_TYPE, 'query'),
+            ...termsQuery(
+              '_id',
+              ids.map((id) => getQueryLinkUuid(name, { [ASSET_TYPE]: 'query', [ASSET_ID]: id }))
+            ),
+          ],
+        },
+      },
+    });
+
+    return assetsResponse.hits.hits.map((hit) => fromStorage(hit._source));
+  }
+
+  async findQueries(name: string, query: string) {
+    const assetsResponse = await this.dependencies.storageClient.search({
+      size: 10_000,
+      track_total_hits: false,
+      runtime_mappings: {
+        [QUERY_FEATURE_NAME]: { type: 'keyword' },
+        [QUERY_FEATURE_FILTER]: { type: 'keyword' },
+      },
+      query: {
+        bool: {
+          filter: [...termQuery(STREAM_NAME, name), ...termQuery(ASSET_TYPE, 'query')],
+          should: [
+            ...wildcardQuery(QUERY_TITLE, query),
+            ...wildcardQuery(QUERY_KQL_BODY, query),
+            ...wildcardQuery(QUERY_FEATURE_NAME, query),
+            ...wildcardQuery(QUERY_FEATURE_FILTER, query),
+          ],
+          minimum_should_match: 1,
+        },
+      },
+    });
+
+    return assetsResponse.hits.hits.map((hit) => fromStorage(hit._source));
+  }
+
+  private async bulkStorage(name: string, operations: QueryBulkOperation[]) {
+    return await this.dependencies.storageClient.bulk({
+      operations: operations.map((operation) => {
+        if ('index' in operation) {
+          const document = toStorage(name, Object.values(operation)[0].asset as AssetLinkRequest);
+          return {
+            index: {
+              document,
+              _id: document[ASSET_UUID],
+            },
+          };
+        }
+
+        const id = getQueryLinkUuid(name, operation.delete.asset);
+        return {
+          delete: {
+            _id: id,
+          },
+        };
+      }),
+      throwOnFail: true,
+    });
+  }
+
+  async getAssets(name: string): Promise<Asset[]> {
+    const { [name]: queryLinks } = await this.getQueryLinks([name]);
+
+    if (queryLinks.length === 0) {
+      return [];
+    }
+
+    return queryLinks.map((link) => {
+      return {
+        [ASSET_ID]: link[ASSET_ID],
+        [ASSET_UUID]: link[ASSET_UUID],
+        [ASSET_TYPE]: link[ASSET_TYPE],
+        query: link.query,
+        title: link.query.title,
+      };
+    });
+  }
+
+  // ==================== Query Sync with Rules ====================
 
   public async syncQueries(stream: string, queries: StreamQuery[]) {
     if (!this.isSignificantEventsEnabled) {
@@ -62,16 +397,13 @@ export class QueryClient {
      * - If a query is updated without a breaking change, it updates the existing rule.
      * - If a query is deleted, it removes the associated rule.
      */
-    const { [stream]: currentQueryLinks } = await this.dependencies.assetClient.getAssetLinks(
-      [stream],
-      ['query']
-    );
+    const { [stream]: currentQueryLinks } = await this.getQueryLinks([stream]);
     const currentIds = new Set(currentQueryLinks.map((link) => link.query.id));
     const nextIds = new Set(queries.map((query) => query.id));
 
     const nextQueriesToCreate = queries
       .filter((query) => !currentIds.has(query.id))
-      .map((query) => toQueryLink(query, stream));
+      .map((query) => toQueryLinkFromQuery(query, stream));
 
     const currentQueriesToDelete = currentQueryLinks.filter((link) => !nextIds.has(link.query.id));
 
@@ -87,7 +419,7 @@ export class QueryClient {
           return hasBreakingChange(currentLink!.query, query);
         }
       ),
-      (partitionedQueries) => partitionedQueries.map((query) => toQueryLink(query, stream))
+      (partitionedQueries) => partitionedQueries.map((query) => toQueryLinkFromQuery(query, stream))
     );
 
     await this.uninstallQueries([...currentQueriesToDelete, ...currentQueriesToDeleteBeforeUpdate]);
@@ -97,14 +429,13 @@ export class QueryClient {
       stream
     );
 
-    await this.dependencies.assetClient.syncAssetList(
+    await this.syncQueryList(
       stream,
       queries.map((query) => ({
         [ASSET_ID]: query.id,
         [ASSET_TYPE]: 'query',
         query,
-      })),
-      'query'
+      }))
     );
   }
 
@@ -138,10 +469,7 @@ export class QueryClient {
       return;
     }
 
-    const { [stream]: currentQueryLinks } = await this.dependencies.assetClient.getAssetLinks(
-      [stream],
-      ['query']
-    );
+    const { [stream]: currentQueryLinks } = await this.getQueryLinks([stream]);
     const queriesToDelete = currentQueryLinks.map((link) => ({ delete: { id: link.query.id } }));
     await this.bulk(stream, queriesToDelete);
   }
@@ -157,10 +485,7 @@ export class QueryClient {
       return;
     }
 
-    const { [stream]: currentQueryLinks } = await this.dependencies.assetClient.getAssetLinks(
-      [stream],
-      ['query']
-    );
+    const { [stream]: currentQueryLinks } = await this.getQueryLinks([stream]);
     const currentIds = new Set(currentQueryLinks.map((link) => link.query.id));
     const indexOperationsMap = new Map(
       operations
@@ -180,7 +505,7 @@ export class QueryClient {
         }),
       ...operations
         .filter((operation) => operation.index && !currentIds.has(operation.index!.id))
-        .map((operation) => toQueryLink(operation.index!, stream)),
+        .map((operation) => toQueryLinkFromQuery(operation.index!, stream)),
     ];
 
     await this.syncQueries(
