+++ x-pack/platform/plugins/shared/streams/server/lib/streams/state_management/types.ts
@@ -8,7 +8,6 @@
 import type { IScopedClusterClient, Logger } from '@kbn/core/server';
 import type { Streams } from '@kbn/streams-schema';
 import type { LockManagerService } from '@kbn/lock-manager';
-import type { AssetClient } from '../assets/asset_client';
 import type { StreamsClient } from '../client';
 import type { StreamsStorageClient } from '../storage/streams_storage_client';
 import type { QueryClient } from '../assets/query/query_client';
@@ -33,7 +32,6 @@ export interface StateDependencies {
   streamsClient: StreamsClient;
   storageClient: StreamsStorageClient;
   scopedClusterClient: IScopedClusterClient;
-  assetClient: AssetClient;
   attachmentClient: AttachmentClient;
   featureClient: FeatureClient;
   queryClient: QueryClient;
