+++ x-pack/platform/plugins/shared/streams/server/plugin.ts
@@ -30,7 +30,6 @@ import {
 import { registerFeatureFlags } from './feature_flags';
 import { ContentService } from './lib/content/content_service';
 import { registerRules } from './lib/rules/register_rules';
-import { AssetService } from './lib/streams/assets/asset_service';
 import { AttachmentService } from './lib/streams/attachments/attachment_service';
 import { QueryService } from './lib/streams/assets/query/query_service';
 import { StreamsService } from './lib/streams/service';
@@ -107,7 +106,6 @@ export class StreamsPlugin
     registerRules({ plugins, logger: this.logger.get('rules') });
     registerStreamsSavedObjects(core.savedObjects);
 
-    const assetService = new AssetService(core, this.logger);
     const attachmentService = new AttachmentService(core, this.logger);
     const streamsService = new StreamsService(core, this.logger, this.isDev);
     const featureService = new FeatureService(core, this.logger, getDefaultFeatureRegistry());
@@ -122,25 +120,21 @@ export class StreamsPlugin
     }): Promise<RouteHandlerScopedClients> => {
       const [
         [coreStart, pluginsStart],
-        assetClient,
         attachmentClient,
         featureClient,
         contentClient,
+        queryClient,
       ] = await Promise.all([
         core.getStartServices(),
-        assetService.getClientWithRequest({ request }),
         attachmentService.getClientWithRequest({ request }),
         featureService.getClientWithRequest({ request }),
         contentService.getClient(),
+        queryService.getClientWithRequest({ request }),
       ]);
 
-      const [queryClient, uiSettingsClient] = await Promise.all([
-        queryService.getClientWithRequest({
-          request,
-          assetClient,
-        }),
-        coreStart.uiSettings.asScopedToClient(coreStart.savedObjects.getScopedClient(request)),
-      ]);
+      const uiSettingsClient = coreStart.uiSettings.asScopedToClient(
+        coreStart.savedObjects.getScopedClient(request)
+      );
 
       const scopedClusterClient = coreStart.elasticsearch.client.asScoped(request);
       const soClient = coreStart.savedObjects.getScopedClient(request);
@@ -155,7 +149,6 @@ export class StreamsPlugin
 
       const streamsClient = await streamsService.getClientWithRequest({
         request,
-        assetClient,
         attachmentClient,
         queryClient,
         featureClient,
@@ -164,7 +157,6 @@ export class StreamsPlugin
       return {
         scopedClusterClient,
         soClient,
-        assetClient,
         attachmentClient,
         streamsClient,
         featureClient,
@@ -232,7 +224,6 @@ export class StreamsPlugin
     registerRoutes({
       repository: streamsRouteRepository,
       dependencies: {
-        assets: assetService,
         features: featureService,
         server: this.server,
         telemetry: this.ebtTelemetryService.getClient(),
