+++ x-pack/platform/plugins/shared/streams/server/routes/streams/crud/read_stream.ts
@@ -19,7 +19,7 @@ import type {
   WiredIngestStreamEffectiveFailureStore,
 } from '@kbn/streams-schema/src/models/ingest/failure_store';
 import type { AttachmentClient } from '../../../lib/streams/attachments/attachment_client';
-import type { AssetClient } from '../../../lib/streams/assets/asset_client';
+import type { QueryClient } from '../../../lib/streams/assets/query/query_client';
 import type { StreamsClient } from '../../../lib/streams/client';
 import {
   getDataStreamLifecycle,
@@ -28,25 +28,23 @@ import {
   getUnmanagedElasticsearchAssets,
 } from '../../../lib/streams/stream_crud';
 import { addAliasesForNamespacedFields } from '../../../lib/streams/component_templates/logs_layer';
-import type { QueryLink } from '../../../../common/assets';
-import { ASSET_TYPE } from '../../../lib/streams/assets/fields';
 
 export async function readStream({
   name,
-  assetClient,
+  queryClient,
   attachmentClient,
   streamsClient,
   scopedClusterClient,
 }: {
   name: string;
-  assetClient: AssetClient;
+  queryClient: QueryClient;
   attachmentClient: AttachmentClient;
   streamsClient: StreamsClient;
   scopedClusterClient: IScopedClusterClient;
 }): Promise<Streams.all.GetResponse> {
-  const [streamDefinition, { [name]: assets }, attachments] = await Promise.all([
+  const [streamDefinition, { [name]: queryLinks }, attachments] = await Promise.all([
     streamsClient.getStream(name),
-    assetClient.getAssetLinks([name], ['query']),
+    queryClient.getQueryLinks([name]),
     attachmentClient.getAttachments(name),
   ]);
 
@@ -62,20 +60,7 @@ export async function readStream({
     { dashboards: [] as string[], rules: [] as string[] }
   );
 
-  const assetsByType = assets.reduce(
-    (acc, asset) => {
-      const assetType = asset[ASSET_TYPE];
-      if (assetType === 'query') {
-        acc.queries.push(asset);
-      }
-      return acc;
-    },
-    {
-      queries: [] as QueryLink[],
-    }
-  );
-
-  const queries = assetsByType.queries.map((query) => {
+  const queries = queryLinks.map((query) => {
     return query.query;
   });
