+++ x-pack/platform/plugins/shared/streams/server/routes/streams/crud/route.ts
@@ -40,14 +40,14 @@ export const readStreamRoute = createServerRoute({
     getScopedClients,
     server,
   }): Promise<Streams.all.GetResponse> => {
-    const { assetClient, attachmentClient, streamsClient, scopedClusterClient } =
+    const { queryClient, attachmentClient, streamsClient, scopedClusterClient } =
       await getScopedClients({
         request,
       });
 
     const body = await readStream({
       name: params.path.name,
-      assetClient,
+      queryClient,
       attachmentClient,
       scopedClusterClient,
       streamsClient,
