+++ x-pack/platform/plugins/shared/streams/server/routes/types.ts
@@ -14,9 +14,7 @@ import type { DefaultRouteHandlerResources } from '@kbn/server-route-repository'
 import type { IUiSettingsClient } from '@kbn/core/server';
 import type { IFieldsMetadataClient } from '@kbn/fields-metadata-plugin/server/services/fields_metadata/types';
 import type { ContentClient } from '../lib/content/content_client';
-import type { AssetClient } from '../lib/streams/assets/asset_client';
 import type { AttachmentClient } from '../lib/streams/attachments/attachment_client';
-import type { AssetService } from '../lib/streams/assets/asset_service';
 import type { QueryClient } from '../lib/streams/assets/query/query_client';
 import type { StreamsClient } from '../lib/streams/client';
 import type { EbtTelemetryClient } from '../lib/telemetry';
@@ -35,7 +33,6 @@ export type GetScopedClients = ({
 export interface RouteHandlerScopedClients {
   scopedClusterClient: IScopedClusterClient;
   soClient: SavedObjectsClientContract;
-  assetClient: AssetClient;
   attachmentClient: AttachmentClient;
   streamsClient: StreamsClient;
   featureClient: FeatureClient;
@@ -49,7 +46,6 @@ export interface RouteHandlerScopedClients {
 }
 
 export interface RouteDependencies {
-  assets: AssetService;
   server: StreamsServer;
   telemetry: EbtTelemetryClient;
   getScopedClients: GetScopedClients;
